requirements.txt
python-telegram-bot==20.3
requests
openai==0.28
pytz
psycopg2-binary
python-dotenv